Use ts and bun
